requirements.txt
asgiref==3.6.0
attrs==22.2.0
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
cytoolz==0.12.1
Django==3.2.18
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
hexbytes==0.3.0
idna==3.4
importlib-metadata==6.1.0
ipfshttpclient==0.6.1
jsonschema==3.2.0
lru-dict==1.1.8
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
pip==22.3.1
protobuf==3.20.3
pycryptodome==3.17
pypiwin32==223
pyrsistent==0.19.3
pytz==2022.7.1
pywin32==305
requests==2.28.2
rlp==2.0.1
setuptools==65.5.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
typing-extensions==3.10.0.2
urllib3==1.26.15
varint==1.0.2
web3==5.12.2
websockets==8.1
wheel==0.38.4
zipp==3.15.0
